refactor: Move TypeScript from devDependencies to dependencies
This commit updates the project to use TypeScript as a dependency instead of a devDependency. The TypeScript package has been removed from the devDependencies list and added to the dependencies list in the package.json file.

This change was made to ensure that the project has access to the TypeScript package in all environments and to avoid any issues that could arise from using devDependencies in production.

To apply this change, run the following command:

npm install typescript --save

Note: If you are running this command in a development environment, you may need to run `npm install` with the `--production=false` flag to install both production and development dependencies.